refactor: extract pane size management into usePaneSizes hook
* Created a reusable `usePaneSizes` custom hook in `src/hooks/` to manage the local storage of left pane sizes.
* Refactored `BookReader.tsx` and `Bookshelf.tsx` to use the new hook, eliminating completely duplicated logic.
