+++ docs/ja_JP/CHANGELOG.md
@@ -25,7 +25,7 @@
 * コンポーネントの遅延読み込み（`React.lazy`）に伴う `<Suspense>` の欠落が原因でアプリがクラッシュする可能性を修正し、静的インポートへの切り替えによって初期表示時のUXを改善 (#185)
 * 前回開いた本を復元する初期化処理の際、`getRecentlyReadBooks` API が意図せず重複して呼び出されていた不具合を修正 (#187)
 * EPUB小説を読んだ直後に大容量の漫画ファイルを開くと、メモリ使用量が急増してアプリが強制終了する問題を修正 (#188)
-
+* 画面左側のペイン（サイドバー）のサイズを保存・復元するロジックを、共通のカスタムフック `usePaneSizes` として切り出し (#189)
 
 ## [2.0.2] - 2026-04-15
